Added the Examinable Class code
This is not perfect. Currently we are polling and this will have to be replaced.

Furthermore, the relative locations for examination states are hard coded and will have to be fine tuned in the actual game map with actual game props.

Also need to think about what to do with the box trigger which is currently useless.